Working with com0com by delegate; no Arduino
