+++ NCalcScripts/F8.ini
@@ -1,12 +1,15 @@
 [ExportProperty]
-name='F8com'
+name='Fake8com'
 value='COM2'
 
 [ExportProperty]
-name='F8pill'
+name='Fake8pill'
 value='COM3'
 
 [ExportProperty]
-name='F8parms'
-value='s0,s1,s2,s3,s4,s5,s6,s7'
+name='Fake8parms'
+value='f0,f1,f2,f3,f4,f5,f6'
 
+[ExportProperty]
+name='Fake8rcv'
+value='Arduino'
